requirements.txt
flask~=2.0.0
psycopg2~=2.9.2
python-dotenv-0.19.2